Update apply on boot for O
