+++ app/src/main/java/com/grarak/kerneladiutor/services/profile/Tasker.java
@@ -27,7 +27,7 @@ import android.util.Log;
 import android.widget.Toast;
 
 import com.grarak.kerneladiutor.R;
-import com.grarak.kerneladiutor.services.boot.Service;
+import com.grarak.kerneladiutor.services.boot.ApplyOnBoot;
 import com.grarak.kerneladiutor.utils.Prefs;
 import com.grarak.kerneladiutor.utils.Utils;
 import com.grarak.kerneladiutor.utils.kernel.cpu.CPUFreq;
@@ -73,7 +73,7 @@ public class Tasker extends BroadcastReceiver {
                             CPUFreq.ApplyCpu applyCpu;
                             if (cs[i].startsWith("#") && (applyCpu =
                                     new CPUFreq.ApplyCpu(cs[i].substring(1))).toString() != null) {
-                                for (String applyCpuCommand : Service.getApplyCpu(applyCpu, su)) {
+                                for (String applyCpuCommand : ApplyOnBoot.getApplyCpu(applyCpu, su)) {
                                     Log.i(TAG + ": " + getClass().getSimpleName(), "Run: " + applyCpuCommand);
                                     su.runCommand(applyCpuCommand);
                                 }
