+++ app/src/main/java/com/grarak/kerneladiutor/services/profile/Widget.java
@@ -31,7 +31,7 @@ import android.widget.RemoteViewsService;
 
 import com.grarak.kerneladiutor.R;
 import com.grarak.kerneladiutor.database.tools.profiles.Profiles;
-import com.grarak.kerneladiutor.services.boot.Service;
+import com.grarak.kerneladiutor.services.boot.ApplyOnBoot;
 import com.grarak.kerneladiutor.utils.Prefs;
 import com.grarak.kerneladiutor.utils.Utils;
 import com.grarak.kerneladiutor.utils.kernel.cpu.CPUFreq;
@@ -107,7 +107,7 @@ public class Widget extends AppWidgetProvider {
                         if (command.getCommand().startsWith("#")
                                 && (applyCpu = new CPUFreq.ApplyCpu(command.getCommand()
                                 .substring(1))).toString() != null) {
-                            adjustedCommands.addAll(Service.getApplyCpu(applyCpu, su));
+                            adjustedCommands.addAll(ApplyOnBoot.getApplyCpu(applyCpu, su));
                         } else {
                             adjustedCommands.add(command.getCommand());
                         }
